backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
Django==5.2
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
numpy==2.4.0
openai==2.14.0
openpyxl==3.1.5
pandas==2.3.3
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
